+++ src/main/kotlin/io/github/nchaugen/tabletest/inspections/ConverterAnnotationUtil.kt
@@ -5,7 +5,7 @@ import com.intellij.psi.PsiMethod
 import com.intellij.psi.util.PsiTreeUtil
 
 internal object ConverterAnnotationUtil {
-    internal const val CONVERTER_ANNOTATION = "org.tabletest.junit.Converter"
+    internal const val CONVERTER_ANNOTATION = "org.tabletest.junit.TypeConverter"
 
     fun isConverterAnnotatedJava(element: PsiElement): Boolean {
         val method = PsiTreeUtil.getParentOfType(element, PsiMethod::class.java, false) ?: return false