+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Treat `@Converter` methods as used in both Java and Kotlin by the unused declaration/symbol inspections
+- Treat `@TypeConverter` methods as used in both Java and Kotlin by the unused declaration/symbol inspections
 
 ## [0.2.0] - 2026-01-17