+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.2.1] - 2026-02-05
+
 ### Added
 
 - Support `org.tabletest.junit.TableTest` annotations
@@ -109,7 +111,8 @@
 - Syntax highlighting in TableTest tables
 - Auto-injection of TableTest language for Java `@TableTest` annotations
 
-[Unreleased]: https://github.com/nchaugen/tabletest-intellij/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/nchaugen/tabletest-intellij/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/nchaugen/tabletest-intellij/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/nchaugen/tabletest-intellij/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/nchaugen/tabletest-intellij/compare/v0.0.10...v0.1.0
 [0.0.10]: https://github.com/nchaugen/tabletest-intellij/compare/v0.0.9...v0.0.10